lfg some progress on serialize_param macro

+++ x/programs/examples/imports/pstate/pstate.go
@@ -60,7 +60,7 @@ func (i *Import) Register(link runtime.Link, meter runtime.Meter, _ runtime.Supp
 	return nil
 }
 
-func (i *Import) putFn(caller *wasmtime.Caller, idPtr int64, keyPtr int64, valuePtr int64) int32 {
+func (i *Import) putFn(caller *wasmtime.Caller, idPtr int64, keyPtr int32, valuePtr int32) int32 {
 	client := runtime.NewExportClient(caller)
 	memory := runtime.NewMemory(client)
 	programIDBytes, err := memory.Range(uint64(idPtr), uint64(ids.IDLen))
@@ -71,7 +71,7 @@ func (i *Import) putFn(caller *wasmtime.Caller, idPtr int64, keyPtr int64, value
 		return -1
 	}
 
-	keyBytes, err := imports.GetBytesFromPtr(client, keyPtr)
+	keyBytes, err := imports.GetBytesFromPtr(client, int64(keyPtr))
 	if err != nil {
 		i.log.Error("failed to read key from memory",
 			zap.Error(err),
@@ -79,7 +79,7 @@ func (i *Import) putFn(caller *wasmtime.Caller, idPtr int64, keyPtr int64, value
 		return -1
 	}
 
-	valueBytes, err := imports.GetBytesFromPtr(client, valuePtr)
+	valueBytes, err := imports.GetBytesFromPtr(client, int64(valuePtr))
 	if err != nil {
 		i.log.Error("failed to read value from memory",
 			zap.Error(err),
@@ -99,7 +99,7 @@ func (i *Import) putFn(caller *wasmtime.Caller, idPtr int64, keyPtr int64, value
 	return 0
 }
 
-func (i *Import) getFn(caller *wasmtime.Caller, idPtr int64, keyPtr int64) int64 {
+func (i *Import) getFn(caller *wasmtime.Caller, idPtr int64, keyPtr int32) int64 {
 	client := runtime.NewExportClient(caller)
 	memory := runtime.NewMemory(client)
 	programIDBytes, err := memory.Range(uint64(idPtr), uint64(ids.IDLen))
@@ -110,7 +110,7 @@ func (i *Import) getFn(caller *wasmtime.Caller, idPtr int64, keyPtr int64) int64
 		return -1
 	}
 
-	keyBytes, err := imports.GetBytesFromPtr(client, keyPtr)
+	keyBytes, err := imports.GetBytesFromPtr(client, int64(keyPtr))
 	if err != nil {
 		i.log.Error("failed to read key from memory",
 			zap.Error(err),
@@ -135,7 +135,7 @@ func (i *Import) getFn(caller *wasmtime.Caller, idPtr int64, keyPtr int64) int64
 		)
 		return -1
 	}
-	
+
 	// prepend the length so that the program can grab the correct number of bytes
 	val = imports.PrependLength(val)
 	ptr, err := runtime.WriteBytes(memory, val)
